requirements.txt
google-generativeai>=0.8.0
python-dotenv>=1.0.0
numpy<2

# Semantic memory (optional but recommended)
chromadb>=0.5.0
sentence-transformers>=3.0.0
